+++ docs/releases/releases.md
@@ -33,8 +33,6 @@ The main changes in this release are:
 
 **Added**
 
-+ Support for `Python 3.14`, deprecated support for `Python 3.9`.
-
 + New class <code>[ForecasterRecursiveClassifier]</code> in the <code>[recursive]</code> module. This forecaster is designed to handle time series data where the target variable is categorical, enabling the prediction of future class labels based on historical patterns.
 
 + New class <code>[PopulationDriftDetector]</code> in the <code>[drift_detection]</code> module to detect population drift between reference and new data. Suitable to detect when forecasting models need to be retrained due to changes in the data distribution. It supports both target and exogenous variables, in single and multiseries forecasting.
@@ -50,11 +48,14 @@ The main changes in this release are:
 
 **Changed**
 
++ Deprecated support for `Python 3.9`.
+
 + Parameter `regressor` has been deprecated in favor of `estimator` in all Forecasters and will be removed in future releases to align with scikit-learn terminology. Visit the [migration guide](../user_guides/migration-guide.html) section for more information.
 
 + Class <code>[Sarimax]</code> has been moved to the new <code>[stats]</code> module. Visit the [migration guide](../user_guides/migration-guide.html) section for more information.
 
 + Class <code>[ForecasterSarimax]</code> has been deprecated in favor of the new <code>[ForecasterStats]</code> model in the <code>[recursive]</code> module. The new forecaster is compatible with a broader range of statistical models such as: sarimax, arima, arar and ets. Visit the [migration guide](../user_guides/migration-guide.html) section for more information.
+
 
 **Fixed**